import tool tips - move alerting to bif
